Moved cps and scm-step tests.
