+++ cps/src/test/resources/org/jenkinsci/plugins/workflow/cps/steps/localMethodCallWithinLotsOfBranches.groovy
