+++ pom.xml
@@ -60,9 +60,9 @@
         <module>durable-task-step</module>
         <module>scm-step</module>
         <module>cps</module>
-        <module>cps-global-lib</module>
         <module>job</module>
         <module>multibranch</module>
+        <module>cps-global-lib</module>
         <module>aggregator</module>
     </modules>
     <dependencyManagement> <!-- TODO perhaps inline -->
@@ -81,21 +81,6 @@
                 <groupId>org.jenkins-ci.plugins</groupId>
                 <artifactId>scm-api</artifactId>
                 <version>1.1</version>
-            </dependency>
-            <dependency>
-                <groupId>org.jenkins-ci.plugins</groupId>
-                <artifactId>git</artifactId>
-                <version>2.4.1</version>
-            </dependency>
-            <dependency>
-                <groupId>org.jenkins-ci.plugins</groupId>
-                <artifactId>subversion</artifactId>
-                <version>2.5</version>
-            </dependency>
-            <dependency>
-                <groupId>org.jenkins-ci.plugins</groupId>
-                <artifactId>mercurial</artifactId>
-                <version>1.54</version>
             </dependency>
             <dependency>
                 <groupId>org.mockito</groupId>
