+++ scm-step/src/test/java/org/jenkinsci/plugins/workflow/steps/scm/AbstractSampleRepoRule.java
